watchdog_be/pyproject.toml
[tool.poetry]
name = "watchdog_be"
version = "0.1.0"
description = ""
authors = ["Ilman Nafian <milmannafian04@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8.5,<3.9"
Django = "^3.1.2"
django-environ = "^0.4.5"
gunicorn = "^20.0.4"
daphne = "^2.5.0"
channels = "^2.4.0"
channels-redis = "^3.1.0"
psycopg2 = "^2.8.6"
django-cors-headers = "^3.5.0"
djangorestframework = "^3.12.1"
djangorestframework-simplejwt = "^4.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
